+++ src/store/index.ts
@@ -1,9 +1,0 @@
-export { alarmModule } from '@/store/modules/alarms';
-export { lightModule } from '@/store/modules/light';
-export { loadingModule } from '@/store/modules/loading';
-export { piStatusModule } from '@/store/modules/piStatus';
-export { settingsModule } from '@/store/modules/settings';
-export { snackbarModule } from '@/store/modules/snackbar';
-export { userModule } from '@/store/modules/user';
-export { pagetitleModule } from '@/store/modules/pagetitle';
-export { wsModule } from '@/store/modules/websocket';