+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# <a href='https://github.com/mrjackwills/belugasnooze_vue/releases/tag/v0.2.5'>v0.2.5</a>
+### 2023-05-11
+
+### Chores
++ dependencies updated, [e5e4695c](https://github.com/mrjackwills/belugasnooze_vue/commit/e5e4695c1e9199957e10e010fbf7d762dc59c581), [af020ea1](https://github.com/mrjackwills/belugasnooze_vue/commit/af020ea1a430a3a1274224a18cace392b5112f0e)
+
+### Fixes
++ invalid types, [bcec20f8](https://github.com/mrjackwills/belugasnooze_vue/commit/bcec20f8f1413d9cc076f9a4cf2488d588319651)
++ 100dvh wrap removed, [a16029b3](https://github.com/mrjackwills/belugasnooze_vue/commit/a16029b38cefe646386575522ede97b13646f7a8)
++ autoimport pinia modules, [55b35ecf](https://github.com/mrjackwills/belugasnooze_vue/commit/55b35ecfb838220558c497272f2dde99425729b5), [1d2eaa14](https://github.com/mrjackwills/belugasnooze_vue/commit/1d2eaa14d97a116f6c019b5bf617bdbce83e855a)
+
+### Reverts
++ temp snackbar fix removed, [7867d3a3](https://github.com/mrjackwills/belugasnooze_vue/commit/7867d3a337bdc4ce5d58c61ac5564ef5f1d5f87a)
+
 # <a href='https://github.com/mrjackwills/belugasnooze_vue/releases/tag/v0.2.4'>v0.2.4</a>
 ### 2023-03-10